feat: Show filename when importing with PartSeg in napari (#1226)
Hi,

Thank you for developing this fantastic tool! It has been incredibly
valuable for our research.

While working with PartSeg in Napari, we noticed that the filename does
not appear in the layer names, making distinguishing between multiple
imported files challenging. To address this, I’ve updated the code in
this pull request to add the filename as a prefix to the layer names.

I’m open to feedback on this approach and would love your thoughts or
suggestions for further improvement!

Best,
David

## Summary by Sourcery

New Features:
- Add filename as a prefix to layer names when importing with PartSeg in
Napari.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced a new method for adjusting color values, enhancing color
processing capabilities.
- Added conditional functionality for adding color based on the version
of the application.
- Improved layer naming conventions by incorporating file name prefixes
for better clarity.
	- Added a new property for selecting layer naming formats in settings.
- Introduced a new enumeration for layer naming formats with multiple
options.
	- Added new test cases to validate layer naming format functionality.

- **Bug Fixes**
- Enhanced handling of alpha channels in color strings to ensure
accurate color representation.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>